requirements.txt
colorlog==6.8.2
hid==1.0.6
mutagen==1.47.0
PyYAML==6.0.2
tqdm==4.66.5
